chore: move stateless transaction validator config to config file
